+++ DEV_ONLY/Example.js
@@ -2,6 +2,7 @@ import React, {
   PureComponent
 } from 'react';
 import WindowedList from '../src';
+import uuid from 'uuid/v4';
 
 const renderItem = (index, key) => {
   return (
@@ -191,6 +192,43 @@ const hiddenProps = {
   type: 'uniform'
 };
 
+const differentKeyItems = (new Array(10000))
+  .fill('foo')
+  .map(() => {
+    return {
+      id: uuid(),
+      height: 100 + Math.floor(Math.random() * 10)
+    };
+  });
+
+const differentKeyRenderer = (index) => {
+  const item = differentKeyItems[index];
+  const style = {
+    height: item.height
+  };
+
+  return (
+    <div
+      className={`item${index % 2 ? '' : ' even'}`}
+      key={item.id}
+      style={style}
+    >
+      {index}
+    </div>
+  );
+};
+
+differentKeyRenderer.toJSON = () => {
+  return differentKeyRenderer.toString();
+};
+
+const differentKeyProps = {
+  length: 10000,
+  itemRenderer: differentKeyRenderer,
+  threshold: 400,
+  type: 'variable'
+};
+
 class Example extends PureComponent {
   state = {
     isVisible: false
@@ -245,6 +283,26 @@ class Example extends PureComponent {
             >
               <WindowedList
                 {...hiddenProps}
+              />
+            </div>
+          </div>
+
+          <div className="example axis-y">
+            <strong>
+              Props
+            </strong>
+
+            <pre className="props">
+              {JSON.stringify(differentKeyProps, null, 2)}
+            </pre>
+
+            <strong>
+              Component
+            </strong>
+
+            <div className="component">
+              <WindowedList
+                {...differentKeyProps}
               />
             </div>
           </div>