+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # react-windowed-list CHANGELOG
 
+### 1.2.1
+* Fix jitter issue when an alternative key is used for list items when type is `variable`
+
+### 1.2.0
+* Make lazy loading opt-in via `isLazy` prop
+
 ### 1.1.0
 * Add `debounceReconciler` prop to allow ability to debounce post-update reconciliations for certain edge cases